+++ lib/data/models/form_data.dart
@@ -1,0 +1,15 @@
+class FormDataFieldModel {
+  final String name;
+  final String value;
+
+  FormDataFieldModel(this.name, this.value);
+}
+
+
+class FormDataFileModel {
+  final String? fileName;
+  final String contentType;
+  final int length;
+
+  FormDataFileModel(this.fileName, this.contentType, this.length);
+}