+++ lib/data/repositories/dio_request_repository_impl.dart
@@ -1,4 +1,5 @@
 import 'package:dio_request_inspector/data/datasources/local_data_source.dart';
+import 'package:dio_request_inspector/data/mappers/mapper.dart';
 import 'package:dio_request_inspector/data/models/http_activity.dart';
 import 'package:dio_request_inspector/domain/repositories/dio_request_repository.dart';
 import 'package:dartz/dartz.dart';
@@ -12,33 +13,51 @@ class DioRequestRepositoryImpl implements DioRequestRepository {
 
   @override
   Future<Either<Failure, Stream<List<HttpActivity>>>> getListResponse() async {
-    final response = localDataSource!.getAllResponse;
-    return Right(response);
+    try {
+      final response = localDataSource!.getAllResponse;
+      return Right(response);
+    } catch (e) {
+      return Left(GeneralFailure(e.toString()));
+    }
   }
 
   @override
   Future<Either<Failure, String>> saveResponse(Response response) async {
-    final result = await localDataSource?.saveResponse(response);
-    return Right(result!);
+    try {
+      final result = await localDataSource!.saveResponse(response.toHttpResponse());
+      return Right(result);
+    } catch (e) {
+      return Left(GeneralFailure(e.toString()));
+    }
   }
 
   @override
-  Future<Either<Failure, String>> saveRequestUseCase(
-      RequestOptions options) async {
-    final result = await localDataSource?.saveRequest(options);
-    return Right(result!);
+  Future<Either<Failure, String>> saveRequest(RequestOptions options) async {
+    try {
+      final result = await localDataSource!.saveRequest(options.toHttpRequest());
+      return Right(result);
+    } catch (e) {
+      return Left(GeneralFailure(e.toString()));
+    }
   }
 
   @override
-  Future<Either<Failure, String>> saveErrorUseCase(DioError error) async {
-    final result = await localDataSource?.saveError(error);
-    return Right(result!);
+  Future<Either<Failure, String>> saveError(DioError e) async {
+    try {
+      final result = await localDataSource!.saveError(e.toHttpError(), e.toHttpErrorFromResponse());
+      return Right(result);
+    } catch (e) {
+      return Left(GeneralFailure(e.toString()));
+    }
   }
-  
+
   @override
   Future<Either<Failure, Stream<List<HttpActivity>>>> clearAllLog() async {
-    final response = localDataSource!.clearAllLog();
-    return Right(response);
+    try {
+      final result = localDataSource!.clearAllLog();
+      return Right(result);
+    } catch (e) {
+      return Left(GeneralFailure(e.toString()));
+    }
   }
-
 }