+++ lib/domain/usecases/save_error_usecase.dart
@@ -9,6 +9,6 @@ class SaveErrorUseCase {
   SaveErrorUseCase(this.repository);
 
   Future<Either<Failure, String>> execute(DioError error) {
-    return repository!.saveErrorUseCase(error);
+    return repository!.saveError(error);
   }
 }